chore: Organise file structure
